use qreal

+++ DDNetLib/include/ddnet/util/utility.hpp
@@ -61,8 +61,8 @@ inline constexpr _UnderlyingType toUnderlying(_EnumType value) noexcept {
     return static_cast<_UnderlyingType>(value);
 }
 
-inline constexpr qint32 floatToFixed(float value) noexcept { 
-    return static_cast<int>(value * static_cast<float>(1 << constants::_fractional_bits)); 
+inline constexpr qint32 floatToFixed(qreal value) noexcept { 
+    return static_cast<int>(value * static_cast<qreal>(1 << constants::_fractional_bits));
 }
 
 inline constexpr float fixedToFloat(qint32 value) noexcept {
